최신 수정사항 반영

+++ fulfillment/src/main/java/com/springboot/fulfillment/data/dto/DeliveryCreateDTO.java
@@ -1,5 +1,7 @@
 package com.springboot.fulfillment.data.dto;
 
+import com.springboot.fulfillment.data.entity.Orders;
+
 import lombok.Getter;
 import lombok.Setter;
 
@@ -9,7 +11,7 @@ public class DeliveryCreateDTO {
 	
 	private String dDeliveryId;
 	
-	private Integer orderId;
+	private Long orderId;
 	
 	private Integer quantity;
